Updated to test using doubles as well
git-svn-id: http://anonsvn.jboss.org/repos/blacktie/trunk@1233 aa7494b7-e75f-0410-880c-f61b6724bf28

+++ jatmibroker-xatmi/src/main/java/org/jboss/blacktie/jatmibroker/xatmi/Buffer.java
@@ -866,7 +866,7 @@ public class Buffer implements Serializable {
 		readPad(dis, currentPos % LONG_SIZE);
 
 		currentPos += LONG_SIZE;
-		double x = dis.readLong();
+		double x = dis.readDouble();
 		ByteBuffer bbuf = ByteBuffer.allocate(LONG_SIZE);
 		bbuf.order(ByteOrder.LITTLE_ENDIAN);
 		bbuf.putDouble(x);
